requirements.txt
alembic==1.0.10
Click==7.0
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Mako==1.0.9
MarkupSafe==1.1.1
mysql-connector-python==8.0.16
protobuf==3.7.1
PyMySQL==0.9.3
SQLAlchemy==1.3.3
Werkzeug==1.0.1
WTForms==2.2.1
